requirements.txt
pymongo==3.12
pandas==1.2.3
memory_profiler==0.58.0
psycopg2==2.8.6
requests==2.24.0
tabulate==0.8.9
python-dotenv==0.15.0
xlsxwriter~=1.3.7
dnspython~=2.1.0
